+++ app/src/main/cpp/opengl/Md2.h
@@ -104,7 +104,7 @@ namespace md2model
         Md2(const char *md2FileName, const char *textureFileName);
         ~Md2();
         // The frame parameter start at 0
-        void Draw(int frame, float xAngle, float yAngle, float scale, float interpolation, glm::mat4 view, glm::mat4 projection);
+        void Draw(size_t frame, float xAngle, float yAngle, float scale, float interpolation, glm::mat4 view, glm::mat4 projection);
 
     private:
         void LoadModel(char *md2FileName);
